Fix display of saving checkbox on the block checkout + moving some files
