+++ client/classic/upe/payment-processing.js
@@ -24,6 +24,7 @@ import {
 	PAYMENT_METHOD_MULTIBANCO,
 	PAYMENT_METHOD_WECHAT_PAY,
 } from 'wcstripe/stripe-utils/constants';
+import { handleDisplayOfSavingCheckbox } from 'wcstripe/classic/upe/spe/handle-display-of-saving-checkbox';
 
 const gatewayUPEComponents = {};
 const paymentMethodsConfig = getStripeServerData()?.paymentMethodsConfig;
@@ -313,6 +314,13 @@ export async function mountStripePaymentElement( api, domElement ) {
 		// Setting the flag to true to prevent the form from being submitted.
 		gatewayUPEComponents[ paymentMethodType ].hasLoadError = true;
 	} );
+
+	// If the SPE is enabled, we need to handle the display of the saving checkbox.
+	if ( getStripeServerData()?.isSPEEnabled ) {
+		upeElement.on( 'change', ( { value } ) => {
+			handleDisplayOfSavingCheckbox( value.type );
+		} );
+	}
 
 	return gatewayUPEComponents[ paymentMethodType ];
 }
