+++ client/classic/upe/spe/__tests__/apply-single-payment-element-styles.test.js
@@ -1,4 +1,4 @@
-import { applySinglePaymentElementStyles } from 'wcstripe/classic/upe/apply-single-payment-element-styles';
+import { applySinglePaymentElementStyles } from 'wcstripe/classic/upe/spe/apply-single-payment-element-styles';
 
 describe( 'applySinglePaymentElementStyles', () => {
 	it( 'Correctly apply the required styles to HTML elements', () => {
