+++ client/classic/upe/spe/apply-single-payment-element-styles.js
